Take ppc64 startup further along the road  - fixed launcher.c to recognise ppc32/64-linux platforms properly  - lots of assembly fixes to handle func descriptors, toc references, 64bit regs.  - fixed var types in vki-ppc64-linux
Now gets as far as VG_(translate), but dies from a case of invalid orig_addr.




git-svn-id: svn://svn.valgrind.org/valgrind/trunk@5299

+++ include/vki-ppc64-linux.h
@@ -399,10 +399,10 @@ struct vki_sigcontext {
 struct vki_stat {
   unsigned long   st_dev;
   unsigned long   st_ino;
-  unsigned short  st_nlink;
-  unsigned short  st_mode;
-  unsigned short  st_uid;
-  unsigned short  st_gid;
+  unsigned long   st_nlink;
+  unsigned int    st_mode;
+  unsigned int    st_uid;
+  unsigned int    st_gid;
   unsigned long   st_rdev;
   long            st_size;
   unsigned long   st_blksize;
